Modified variable names and logic to make this more obvious.
This specific piece of code checks if we were already authenticated before
calling processToken here. In the case that we were already authenticated but
have requested a new token, it doesn't make much sense to throw another
'authenticated' event.

The 'token:changed' event will always be thrown in this case, but no actual
authentication has occured.